+++ src/module.ts
@@ -1,20 +1,31 @@
-import { defineNuxtModule, addPlugin, createResolver, addTypeTemplate } from '@nuxt/kit'
+import {
+  defineNuxtModule,
+  addPlugin,
+  createResolver,
+  addTypeTemplate,
+  addImportsDir,
+  addComponentsDir,
+} from "@nuxt/kit";
 
 export default defineNuxtModule({
-  meta: {
-    name: 'nuxt-toasts',
-    configKey: 'nuxtToasts',
-  },
   defaults: {},
+  meta: {
+    configKey: "nuxtToasts",
+    name: "nuxt-toasts",
+  },
   setup(_options, _nuxt) {
-    const resolver = createResolver(import.meta.url)
+    const resolver = createResolver(import.meta.url);
 
     addTypeTemplate({
-      filename: './types/nuxt-toasts.d.ts',
-      src: resolver.resolve('./types/nuxt-toasts.d.ts'),
-    })
+      filename: "./types/nuxt-toasts.d.ts",
+      src: resolver.resolve("./runtime/types/index.d.ts"),
+    });
+
+    addComponentsDir({ path: resolver.resolve("./runtime/components") });
+    addImportsDir(resolver.resolve("./runtime/components"));
+    addImportsDir(resolver.resolve("./runtime/composables"));
 
     // Do not add the extension since the `.ts` will be transpiled to `.mjs` after `npm run prepack`
-    addPlugin(resolver.resolve('./runtime/plugin'))
+    addPlugin(resolver.resolve("./runtime/plugin"));
   },
-})
+});