+++ src/runtime/composables/useNotificationStore.ts
@@ -1,46 +1,39 @@
-import { createPinia, defineStore } from 'pinia'
-import _ from 'lodash'
-import { ref } from 'vue'
+import { createPinia, defineStore } from "pinia";
+import { ref } from "vue";
 
-interface Notification {
-  id: string
-  title: string
-  text?: string
-  type: 'success' | 'warning' | 'info' | 'error'
-  interval: number
-  callback?: Funcion
-}
+import type { Notification } from "../types";
+import _ from "../packages/lodash.js";
 
-const pinia = createPinia()
+const pinia = createPinia();
 
-export const useNotificationStore = defineStore('notification', () => {
-  const notifications = ref<Notification[]>([])
+export const useNotificationStore = defineStore("notification", () => {
+  const notifications = ref<Notification[]>([]);
 
   const addNotification = (
-    type: 'success' | 'warning' | 'info' | 'error',
+    type: "success" | "warning" | "info" | "error",
     title: string,
-    text: string = '',
+    text: string = "",
     interval: number = 3000,
-    callback: any = null
+    callback: any = null,
   ) => {
-    const notification: Notification = {
+    const notification = {
       id: _.uniqueId(),
       title,
       text,
       type,
       interval,
       callback,
-    }
+    };
 
     // Beschikbare types: success, warning, info en error
-    notifications.value.push(notification)
+    notifications.value.push(notification);
 
     setTimeout(() => {
-      remove(notification.id)
-    }, interval)
+      removeNotification(notification.id);
+    }, interval);
 
-    return notification
-  }
+    return notification;
+  };
 
   /**
    * Adds a notification.
@@ -63,17 +56,17 @@ export const useNotificationStore = defineStore('notification', () => {
    * const params = { type: 'success', title: 'Great!', text: 'Everything is fine.', interval: 2000 };
    * add(params);
    */
-  const add = async (notification: Partial<Omit<Notification, 'id'>> = {}) => {
+  const add = async (notification: Partial<Omit<Notification, "id">> = {}) => {
     const {
-      type = 'info',
-      title = '',
-      text = '',
+      type = "info",
+      title = "",
+      text = "",
       interval = 3000,
       callback = null,
-    } = notification
+    } = notification;
 
-    return addNotification(type, title, text, interval, callback)
-  }
+    return addNotification(type, title, text, interval, callback);
+  };
 
   /**
    * Removes a notification.
@@ -87,15 +80,17 @@ export const useNotificationStore = defineStore('notification', () => {
    * // Removes a notification
    * remove(notificationId);
    */
-  function remove(notificationId: string) {
-    _.remove(notifications.value, { id: notificationId })
+  function removeNotification(notificationId?: string) {
+    _.remove(notifications.value, { id: notificationId });
   }
 
   return {
     notifications,
     add,
-    remove,
-  }
-})
+    removeNotification,
+  };
+});
 
-useNotificationStore(pinia)
+useNotificationStore(pinia);
+
+export default useNotificationStore;