+++ src/runtime/types/index.d.ts
@@ -1,0 +1,8 @@
+export interface Notification {
+  id?: string;
+  title: string;
+  text?: string;
+  type: "success" | "warning" | "info" | "error";
+  interval: number;
+  callback?: Function;
+}