+++ src/types/nuxt-toasts.d.ts
@@ -1,8 +1,0 @@
-type Notification = {
-  id: string
-  type: string
-  title: string
-  text: string | null | undefined
-  interval: number
-  callback: null | Function
-}