+++ Changelog.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [0.2.0] - 2025-03-24
+
+### Added
+
+- Fully functional notification component
+- Playground for testing the notification component
+
+
+### Changed
+
+- Types directory
+- Dependencies
+
+
+### Removed
+
+- npm.rc file
+
+---
+
 
 ## [0.1.1] - 2025-03-24
 
@@ -45,5 +65,6 @@ All notable changes to this project will be documented in this file.
 [//]: # (-)
 
 
+[0.2.0]: https://github.com/TimoCuijpers/nuxt-toasts/releases/tag/0.2.0
 [0.1.1]: https://github.com/TimoCuijpers/nuxt-toasts/releases/tag/0.1.1
 [0.1.0]: https://github.com/TimoCuijpers/nuxt-toasts/releases/tag/0.1.0